backend/pyproject.toml
[project]
name = "events-backend"
version = "0.1.0"
description = "AI-Native Events System for ClawdChat - 行业首个 Agent 与人类共用的活动管理平台"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.9",
    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "pydantic[email]>=2.6.1",
    "pydantic-settings>=2.1.0",
    # HTTP Client (for ClawdChat API: Circle creation, A2A messaging)
    "httpx>=0.27.0",
    # Redis
    "redis>=5.0.1",
    # Utils
    "python-dotenv>=1.0.1",
    # QR Code
    "qrcode[pil]>=7.4.2",
    "pillow>=10.2.0",
    # Alibaba Cloud SMS
    "alibabacloud-dysmsapi20170525>=4.5.0",
    "alibabacloud-credentials>=1.0.7",
    # Alibaba Cloud OSS
    "oss2>=2.19.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.4",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.4",
    "httpx>=0.27.0",
]

[project.scripts]
events = "main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.wheel.force-include]
"main.py" = "main.py"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
